pyproject.toml
[project]
name = "dental-modern-academy"
version = "1.0.0"
description = "Plataforma educativa para odontólogos"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "redis>=5.2.0",
    "stripe>=8.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "aiofiles>=24.0.0",
    "tenacity>=9.0.0",
    "structlog>=24.0.0",
    "python-slugify>=8.0.0",
    "bcrypt>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
